UserForm

+++ base/views.py
@@ -4,6 +4,8 @@ from django.db.models import Q
 from .models import Room,Topic
 from django.contrib import messages
 from django.contrib.auth import authenticate,login,logout
+from django.contrib.auth.forms import UserCreationForm
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from .forms import RoomForm
 # Create your views here.
@@ -13,8 +15,11 @@ rooms=[
     {'id':3,'name':'Fronted developers'}
 ]
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated: #避免輸入login page user時候重複登入
+        return redirect('home')
     if request.method == 'POST':
-        username=request.POST.get('username')
+        username=request.POST.get('username').lower()
         password=request.POST.get('password')
         try:
             user=User.objects.get(username=username)
@@ -26,7 +31,7 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request,"Username Or Password not exist")
-    context= {}
+    context= {'page':page}
     return render(request, 'base/login_register.html',context)
 def logoutUser(request):
     logout(request)
@@ -46,6 +51,7 @@ def room(request,pk):
     context={'room':room}
     return render(request,'base/room.html',context)
 
+@login_required(login_url='login')  #當還沒登入的時候鼓勵人登入
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -55,9 +61,13 @@ def createRoom(request):
             return redirect('home')  #使用urls.py的"name"
     context={'form':form}
     return render(request,'base/room_form.html',context)
+
+@login_required(login_url='login') #確保有user登入下才能update
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)  #form的起始值是這個room的資訊
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
     if request.method == 'POST':
         form=RoomForm(request.POST,instance=room) #告訴她哪個room要更新(instance的部分)
         if form.is_valid():
@@ -65,10 +75,28 @@ def updateRoom(request,pk):
             return redirect('home')
     context={'form':form}
     return render(request,'base/room_form.html',context)
+@login_required(login_url='login') #確保有user登入下才能刪除房間
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
     context={'obj':room}
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here!!")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request,'base/delete.html',context)
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False) #freeze it 要馬上使用 user資訊
+            user.username = user.username.lower()
+            user.save()
+            login(request,user) #註冊完馬上登入
+            return redirect('home')
+        else:
+            messages.error(request,"An error occurred while registration")
+    context={'form':form}
+    return render(request, 'base/login_register.html',context)
